Split join button into new component
