+++ packages/app/src/pages/event/index.js
@@ -2,47 +2,25 @@ import React, { useEffect, useState } from 'react'
 import { useParams } from 'react-router-dom'
 import s from './event-page.css'
 import { combine } from '~style'
-import api from '~api'
 import Map from './map'
 import Organizer from './organizer'
 import Display from '~components/text-editor/display'
 import { Helmet } from 'react-helmet'
+import onStart from './onStart'
+import placeholder from './placeholder.json';
+import JoinEventBtn from '~components/join-event'
 
 // million-ignore
 export default function EventPage(props) {
 	const { id } = useParams();
     const [ data, setData ] = useState({
-		...props.data,
-		date: "Mandag 9. sep",
-		start: "19:00",
-		end: "23:00",
-		location: "Realfagskjelleren, Herman Krags veg 12",
-		description: [{ type:"paragraph", children:[{ text:"Informasjon kommer!" }] }],
-		title: "Ukjent Arrangement",
-		header: "black",
-		type: 0
+		...placeholder,
+		...props.data
 	});
 
 	const link = data.location.replace(' ', '+');
 
-	useEffect(() => {
-		api.get(`/v0/events/get?id=${id}`)
-			.then(r => {
-				const header = `url(/${id}.png)`;
-				const description = JSON.parse(r.event.description);
-
-				setData({
-					...data,
-					...r.event,
-					header: header,
-					description: description,
-				});
-
-				// return r;
-			})
-			// .then(r => console.log("State", data))
-			.catch(console.error)
-	}, [])
+	useEffect(() => onStart(id, data, setData), [])
 
 	return (
 		<div className={s.wrapper}>
@@ -91,11 +69,7 @@ export default function EventPage(props) {
 					</ul>
 					<div className={s.actions}>
 						<h2>Påmelding</h2>
-						<button className={combine(s.register, s.btn)} onClick={() => {
-							alert("TODO")
-						}}>
-							Meld deg på
-						</button>
+						<JoinEventBtn/>
 					</div>
 					<Organizer by={"Delta Linjeforening"}/>
 				</div>
