+++ packages/app/src/style/typography.css
@@ -16,6 +16,8 @@ h2 {
 	font-style: normal;
 	font-weight: 400;
 	font-size: 1.4rem;
+
+	margin-bottom: 0.67rem;
 }
 
 h3 {
